+++ BetterBetterTeleporter/Plugin.cs
@@ -8,11 +8,12 @@ namespace BetterBetterTeleporter;
 
 [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
 [BepInDependency("ainavt.lc.lethalconfig", BepInDependency.DependencyFlags.SoftDependency)]
-[BepInDependency("LethalNetworkAPI")]
+[BepInDependency("com.sigurd.csync", "5.0.1")]
 public class Plugin : BaseUnityPlugin
 {
     internal static new ManualLogSource Logger { get; private set; }
     public static Plugin Instance { get; private set; }
+    public static ModConfig ModConfig;
 
     internal void Awake()
     {
@@ -20,7 +21,7 @@ public class Plugin : BaseUnityPlugin
 
         Instance = this;
         Logger = base.Logger;
-        ModConfig.Init(Config);
+        ModConfig = new ModConfig(Config);
 
         if (Chainloader.PluginInfos.ContainsKey("ainavt.lc.lethalconfig"))
         {